+++ tobbo.kirill/I2/methods.cpp
@@ -6,7 +6,7 @@ void count_of_maxes(const int* arr, size_t size, int& maximum, size_t& count_max
 {
   maximum = std::numeric_limits< int >::min();
   count_maximum = std::numeric_limits< size_t >::min();
-  for (int i = 0; i < size; i++)
+  for (size_t i = 0; i < size; i++)
   {
     if (arr[i] > maximum)
     {
@@ -30,7 +30,7 @@ void shift_part_to_end(int* arr, size_t size, unsigned int start_index, unsigned
     throw std::invalid_argument("Wrong input indexes");
   }
   int shift_size = end_index - start_index;
-  for (int i = 0; i < shift_size; i++)
+  for (size_t i = 0; i < shift_size; i++)
   {
     int tmp = arr[size - i - 1];
     arr[size - i - 1] = arr[end_index - i - 1];
